+++ SplunkOpenTelemetryBackgroundExporter/Sources/SplunkOpenTelemetryBackgroundExporter/Exporters/logs/Binary body support/OTLPBackgroundHTTPLogExporterBinary.swift
@@ -51,13 +51,7 @@ public class OTLPBackgroundHTTPLogExporterBinary: OTLPBackgroundHTTPBaseExporter
 
             // If no endpoint is configured, store in pending folder for later
             if isPendingEndpoint {
-                try diskStorage.insert(
-                    storeData,
-                    forKey: KeyBuilder(
-                        requestId.uuidString,
-                        parrentKeyBuilder: getPendingStorageKey()
-                    )
-                )
+                try storePendingData(storeData, requestId: requestId)
                 return .success
             }